Add a replacement for genrules in macros
Often writing macros is considered less effort than writing
proper rules. Therefore, a lot of macros exists; unfortunately,
in the presence of  external repositories, dependencies on local
tools cannot simply be referred to by a string, as that would be
resolved in the outermost context.

Therefore, add an alternative to genrule that allows an explict
argv, and accepts labels in replacements of paths.

  load("@bazel_skylib//:lib.bzl", "actions")
  def to_upper(name="", src=""):
    tool = Label("//macro:my_tool")
    data = Label("//macro:data.txt")
    input = actions.outer_name(src)
    output = actions.outer_name(name + ".txt")
    actions.generic_action(
      name = name,
      tools = [tool],
      srcs = [input, data],
      outs = [output],
      argv = [tool, input, "magic string", data, output]
    )