Remove unnecessary runfiles declaration of copy_file
`copy_file` unnecessarily declares the copied file as a `runfile` in its `DefaultInfo`, which "pollutes" the runfiles directory of all targets depending on the rule. For example, a `cc_library` with a `copy_file` target in its `hdrs` or `srcs` will also have the copied file as a runfile, which is not correct. Returning the copied file in `files` is sufficient for any depending rule to pick the file up as a runfile if necessary.