{copy,write}_file: Do not add non-executables to default_runfiles
copy_file and write_file currently include the created file in their
runfiles even if it is not executable, which makes every rule depending
on it have the file as a runfile (e.g. a `cc_library` depending on a
copied header file via the hdrs attribute).

In an ideal world, according to
https://docs.bazel.build/versions/main/skylark/rules.html#runfiles-features-to-avoid,
copy_file would not need to specify any runfiles in the `DefaultInfo` it
returns. However, due to the existence of rules with legacy behavior,
this would break compatibility (actually, already with `sh_test` in
skylib's unit tests), see
https://github.com/bazelbuild/bazel/issues/15043.

As a compromise that preserves compatibility with legacy rules but
nevertheless cleans up the runfiles tree of depending rules, use the
`data_runfiles` attribute of `DefaultInfo` if the copied file is not
executable.