Add support for new subpackages module for working with native.subpackages
This adds support for friendly APIs on top of the new native.subpackages
support added to Bazel.  This allows code to use subpackage, if the
Bazel version supports it, test for packages existing and construct
targets based on relative subackages.

This Pull requires support for 'loadingtest' has been added to the
'unittest' module.