Load gazelle plugin module from local subdir in bzlmod mode
Needed to allow building/testing gazelle plugin from bazelci and
generating release tarballs when in bzlmod mode.

Note that the local_path_override will be stripped out in the
tarball release, so users of stable skylib releases will not be
affected.

Users of a live/unreleased version of skylib who depend on it via
git_override will need to add a patch to remove or comment out
the local_path_override. I expect the number of such users to be
tiny, but I have added a comment in MODULE.bazel to guide them.