feat: generate doc_extract targets
Allows bzl_library targets to produce documentation without needing to load stardoc or add additional targets.
Note, it would be better to do this with an aspect that visits the bzl_library graph, but the starlark_doc_extract rule is in Bazel java core, and so we cannot spawn actions using its Java code. It needs a main() entry point.

Use a config_setting to make this opt-in, so we don't generate so many targets all the time.